+++ tests/Common.cpp
@@ -6,12 +6,13 @@
 #include <SDL3/SDL_log.h>
 #include <new>
 
-bool ExampleInit(Context &ctx, float wWidth, float wHeight) {
+bool ExampleInit(Context &ctx, Uint32 wWidth, Uint32 wHeight) {
   if (!SDL_Init(SDL_INIT_VIDEO))
     return false;
   new (&ctx.device) Device{SDL_GPU_SHADERFORMAT_SPIRV, true};
 
-  ctx.window = SDL_CreateWindow("candlewick: examples", wWidth, wHeight, 0);
+  ctx.window =
+      SDL_CreateWindow("candlewick: examples", int(wWidth), int(wHeight), 0);
   if (!SDL_ClaimWindowForGPUDevice(ctx.device, ctx.window)) {
     SDL_Log("Error %s", SDL_GetError());
     return false;
